Fix #1, in progress on #2
Pretty confident about KICK and NICK... minor tweaks may be necessary. Considering #1 closed (tentatively).

The solution for #2 comes as `irc.network`, which is an array containing IRC network information, and then `irc.connect` which takes the network name (main key) as an argument to resolve this (examples given near the end of the file as usual)... seems like a pretty extendable model to work with a GUI... but not quite sitting on this. I think identifiers such as `plain_text`, `signed_tls` and `pinned_tls` might later change when I introduce other substrate protocols...